merge:merge

+++ src/main/java/org/esnack24api/esnack24api/product/domain/Product.java
@@ -16,7 +16,7 @@ public class Product {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY) // autoIncrement로 자동 생성되는 pk
-    private int pno;
+    private Long pno;
 
 
     private int price;
